+++ can_servizi.h
@@ -18,6 +18,6 @@ volatile uint8_t get_servizi_brake();
 volatile bool    get_servizi_apps_plausibility();
 volatile bool    get_servizi_brake_plausibility();
 
-volatile float get_tcs_torque_coefficient();
+volatile uint8_t get_tcs_torque_coefficient();
 
 #endif /* _CAN_SERVIZI_H_ */
